requirements.txt
flask
gunicorn
fastapi
uvicorn[standard]
beautifulsoup4
selenium
webdriver-manager
pandas
lxml
pydantic
uvicorn
psycopg2-binary
google-cloud-storage
cloud-sql-python-connector[pg8000]
langchain-community
google-cloud-aiplatform
python-dotenv
openai 
openai-agents 
langchain-community
tiktoken
langchain
graphviz